refactor
Added drive_by_wire.c and drive_by_wire.h files.
