+++ Src/main.c
@@ -54,6 +54,7 @@ UART_HandleTypeDef huart6;
 DMA_HandleTypeDef hdma_usart2_rx;
 DMA_HandleTypeDef hdma_usart6_rx;
 
+
 /******* User External Variables *********/
 uint16_t speed;
 uint8_t mappedSpeed;
@@ -65,6 +66,7 @@ uint8_t data[10];
 uint8_t brakeStatus;
 uint8_t brakeLevel;
 uint16_t mappedBrakeLevel;
+
 
 /* USER CODE BEGIN PV */
 
@@ -83,137 +85,6 @@ static void MX_TIM6_Init(void);
 static void MX_USART6_UART_Init(void);
 void MX_USB_HOST_Process(void);
 
-/****** User Function Prototypes *********/
-void setBrakeLevel(uint8_t brakeLevel);
-void resetBrakeLevel(uint8_t brakeLevel);       /*** Brake-by-wire ***/
-int getBrakeLevel(uint8_t brakeLevel);
-
-int getSteeringAngle(uint8_t steeringAngle);    /*** Steer-by-wire ***/
-void setSteeringAngle(uint8_t steeringAngle);
-
-int map(int value,
-        int inputMin, int inputMax,
-        int outputMin, int outputMax );
-
-void setSpeed(uint16_t speed);                  /*** Throttle-by-wire ***/
-
-void getValues(uint8_t data[]);                 // Get uart data
-void defaultValues(void);                       // İf not any uart data use default values
-void delay_us (uint16_t us);
-void HAL_UART_RxCpltCallback(UART_HandleTypeDef *huart);
-
-/* USER CODE BEGIN PFP */
-
-/********************************************************************
- * getValues: Takes incoming data from uart as a argument.          *
- *            After sets steeringAngle and speed external variables.*
- *                                                                  *
- *******************************************************************/
-void getValues(uint8_t data[]){
-    // Get data and parse.
-    // Set steeringAngle and speed external variables.
-}
-
-/******************************************************************
- * defaultValues: Sets default value for speed and steeringAngle  *
- *                external variables.                             *
- *                                                                *
- *****************************************************************/
-void defaultValues(void){
-    // Set steeringAngle and speed external variables with default values.
-}
-
-void delay_us (uint16_t us)
-{
-	__HAL_TIM_SET_COUNTER(&htim6,0);  // set the counter value a 0
-	while (__HAL_TIM_GET_COUNTER(&htim6) < us);  // wait for the counter to reach the us input in the parameter
-}
-
-/******************************************************************
- * map: Maps value by input interval to output interval.          *
- *      Takes value, inputMin, inputMax, outputMin, outputMax     *
- *      as a argument. Returns mapped value.                      *
- *                                                                *
- *****************************************************************/
-int map(int value, int inputMin, int inputMax, int outputMin, int outputMax ){
-	return (((((1.0*abs(inputMin-value)) / (abs(inputMax - inputMin))) * (abs(outputMax - outputMin))) + outputMin)) ;
-}
-
-/******************************************************************
- * setSpeed: Sets speed of vehicle. Maps speed argument and       *
- *           sets mappedSpeed value to speed of vehicle.          *
- *                                                                *
- *****************************************************************/
-void setSpeed(uint16_t speed){
-    // Get mapped speed by speed argument
-    // Set speed level of vehicle with PWM.
-}
-
-/********************************************************************
- * getBrakeLevel: Maps brakeLevel argument. Returns mapped brake    *
- *                level.                                            *
- *                                                                  *
- *******************************************************************/
-int getBrakeLevel(uint8_t brakeLevel){
-	return  map(brakeLevel, 0, 10 ,0 , 100);
-}
-
-/********************************************************************
- * setBrakeLevel: Sets brake level of vehicle. Gets mapped brake    *
- *                level and controls brake stepper motor with       *
- *                GPIO.                                             *
- *                                                                  *
- *******************************************************************/
-void setBrakeLevel(uint8_t brakeLevel){
-    // Get mapped brake level
-    // Set brake level of vehicle with controlling stepper motor with GPIO
-}
-
-/********************************************************************
- * resetBrakeLevel: Resets brake level of vehicle to starting       *
- *                  position(stop braking). Gets mapped brake level *
- *                  and controls brake stepper motor with GPIO.     *
- *                                                                  *
- *******************************************************************/
-void resetBrakeLevel(uint8_t brakeLevel){
-    // Get mapped brake level
-    // Stop braking with controlling stepper motor with GPIO
-}
-
-/*
- void setRegenerativeBrake(){
-}
-
-void resetRegenerativeBrake(){
-}
-*/
-
-/*********************************************************************
- * getSteeringAngle: Gets steering angle of vehicle and returns it.  *
- *                  Takes steeringAngle argument and calculates      *
- *                  angle change by comparing with previous steering *
- *                  degree.                                          *
- *                  !Vehicle should start at straight steering angle.*
- *                                                                   *
- ********************************************************************/
-int getSteeringAngle(uint8_t steeringAngle){
-    // Get angle change by comparing steeringAngle with prevAngle value
-    // Return mapped angle value
-	return 1;
-}
-
-/*********************************************************************
- * setSteeringAngle: Gets steering angle of vehicle and returns it.  *
- *                  Takes steeringAngle argument and calculates      *
- *                  angle change by comparing with previous steering *
- *                  degree.                                          *
- *                  !Vehicle should start at straight steering angle.*
- *                                                                   *
- ********************************************************************/
-void setSteeringAngle(uint8_t steeringAngle){
-    // Get mapped steering angle
-    // Control steering stepper motor with GPIO
-}
 
 /* USER CODE END PFP */
